openjdk-6-common: move java2-vm to RRECOMMENDS
Even though a vm is a hard dependency it is set as RRECOMMENDS so a single vm can get uninstalled:

   root@beaglebone:~/java# opkg remove openjdk-6-vm-shark
   No packages removed.
   Collected errors:
    * print_dependents_warning: Package openjdk-6-vm-shark is depended upon by packages:
    * print_dependents_warning:         openjdk-6-java
    * print_dependents_warning: These might cease to work if package openjdk-6-vm-shark is removed.

In the above situation all vms (cacao, zero and shark) were installed, but due to llvm problems[1] shark is currently broken. To fix the java install shark needs to get uninstalled.

In theory this change allows people to break the java installation, but in practice it's already broken due to [1].

[1] http://lists.linuxtogo.org/pipermail/openembedded-devel/2011-December/036595.html

Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>
Signed-off-by: Henning Heinold <heinold@inf.fu-berlin.de>