llvm,openjdk: Turn bitbake parse error into warning on finding unsupported CPU
Sometimes we have meta-java in bblayers but one of the CPUs e.g. mips64
which does not support openjdk errors out on parsing. Thats not OK if
say the image thats being generated does not use java at all. So its
ok to warn and inform about it clearly but it should not be error.

Signed-off-by: Khem Raj <raj.khem@gmail.com>