libmatthew: Resolved erroneous symlink libmatthew.jar by providing ALTJARFILENAMES variable.
The java-library.bbclass appends a jar installation step using the variables
JARFILENAME and ALTJARFILENAMES. The original recipe only set the JARFILENAME
variable, which resulted in an erroneous symlink libmatthew.jar to unix-0.5.jar
to be installed.
By setting the ALTJARFILENAMES variable in the recipe, the installation of the
symlink is resolved and the installation of the jar unix.jar is taken care
of with that.